smb_common: one copy of the code both sides must agree on
The SMB2 server and the SMB2 client VFS module each carried their own
signing, encryption and compression.  Not similar implementations --
identical ones.  Normalising away name prefixes and comment reflow, the
LZ77 decoder differed in 0 lines of logic, decrypt_message in 0, and key
derivation in exactly one thing: which label is used to encrypt and which
to decrypt.  That is a parameter, not a reason for a second copy.

Duplication here is worse than the usual kind because these functions do
not merely resemble each other -- they must produce the SAME BYTES or the
protocol fails.  A drifting mirror produces no compile error and no failed
assertion; it produces a session that negotiates cleanly and then fails
every MAC or tag check, which is among the most expensive bugs to chase.
Writing the client's half hit that twice: a hand-copied KDF missing
MODE=counter would have derived silently wrong keys, and reversing the key
direction fails every AEAD tag while looking like a working handshake.

src/smb_common now holds what belongs to the protocol rather than to
either endpoint, in the same position src/nfs_common holds for NFS --
neither side depends on the other, both depend on this:

  smb2.h          the wire definitions.  These were already shared, badly:
                  the client reached into "server/smb/smb2.h", so the
                  layering violation existed before this component did.
  smb_signing.c   KBKDF, preauth hash, signing-key derivation, the
                  HMAC/CMAC/GMAC algorithms, and signature computation.
  smb_encrypt.c   cipher lookup, key-pair derivation (direction as an
                  argument), the AEAD transform build and parse.
  smb_compress.c  LZ77, LZNT1 and LZ77+Huffman, plus the chained,
                  Pattern_V1 and unchained framings.  This file needed no
                  changes at all -- it already had zero references to
                  server state.

What stays behind is what each side decides for itself.  The server keeps
smb_sign_request.c and smb_secure_send.c: applying a signature to a
chimera_smb_request, and choosing which reply to wrap.  The client keeps
its negotiation policy.  Neither is shared because neither is the same
question on both ends.

Shared code cannot claim either side's logging domain, so it logs under
"smb2" -- the protocol rather than an endpoint.

The moved functions also needed SYMBOL_EXPORT.  They were internal to one
library and are now across a boundary, and the build is
-fvisibility=hidden; on macOS the missing exports did not fail the link,
they bound lazily to NULL and crashed on the first call.

A side effect worth having: the client restricted itself to LZ77 because
mirroring the other two codecs was not worth the duplication.  That limit
was set by copy-paste cost rather than by design, and it is now gone --
LZNT1 and LZ77+Huffman are available to it for free.

Net -406 lines.  Verified beyond "the tests pass": encryption still
interoperates on every dialect (AES-128-CCM at 3.0/3.0.2, AES-128-GCM at
3.1.1), and compression still expands 464 wire bytes to 12,368 with zero
decompressions in the control -- the same numbers as before the move, on
both the plain and compress-then-encrypt paths.
