+++ src/server/smb/smb_internal.h
@@ -18,8 +18,9 @@
 #include "common/logging.h"
 #include "common/macros.h"
 #include "common/misc.h"
-#include "smb2.h"
-#include "smb_encrypt.h"
+#include "smb_common/smb2.h"
+#include "smb_common/smb_encrypt.h"
+#include "smb_secure_send.h"
 #include "smb1.h"
 #include "smb_attr.h"
 #include "smb_session.h"
