+++ src/server/smb/smb_notify.h
@@ -7,7 +7,8 @@
 #include <stdint.h>
 #include <pthread.h>
 #include "vfs/vfs_notify.h"
-#include "smb_encrypt.h"
+#include "smb_common/smb_encrypt.h"
+#include "smb_secure_send.h"
 
 struct chimera_smb_request;
 struct chimera_smb_conn;
