+++ src/server/smb/smb_proc_change_notify.c
@@ -7,7 +7,7 @@
 #include "smb_internal.h"
 #include "smb_procs.h"
 #include "smb_notify.h"
-#include "smb2.h"
+#include "smb_common/smb2.h"
 #include "vfs/vfs_notify.h"
 
 int
