+++ src/server/smb/smb_proc_copyoffload.c
@@ -6,7 +6,7 @@
 
 #include "smb_internal.h"
 #include "smb_procs.h"
-#include "smb2.h"
+#include "smb_common/smb2.h"
 #include "smb_session.h"
 #include "vfs/vfs.h"
 #include "vfs/vfs_procs.h"
