+++ src/server/smb/smb_proc_sparse.c
@@ -4,7 +4,7 @@
 
 #include "smb_internal.h"
 #include "smb_procs.h"
-#include "smb2.h"
+#include "smb_common/smb2.h"
 #include "vfs/vfs.h"
 #include "vfs/vfs_procs.h"
