+++ src/smb_common/smb2.h
