+++ src/smb_common/smb_signing.c
@@ -10,8 +10,8 @@
 #include <openssl/kdf.h>
 
 #include "smb_signing.h"
+#include "smb_common.h"
 #include "common/evpl_iovec_cursor.h"
-#include "smb_internal.h"
 #include "smb2.h"
 
 struct chimera_smb_signing_ctx {
@@ -20,7 +20,7 @@ struct chimera_smb_signing_ctx {
     EVP_CIPHER *gcm;        /* AES-128-GCM, used as GMAC for AES-128-GMAC signing */
 };
 
-struct chimera_smb_signing_ctx *
+SYMBOL_EXPORT struct chimera_smb_signing_ctx *
 chimera_smb_signing_ctx_create(void)
 {
     struct chimera_smb_signing_ctx *ctx = calloc(1, sizeof(struct chimera_smb_signing_ctx));
@@ -31,20 +31,20 @@ chimera_smb_signing_ctx_create(void)
 
     ctx->hmac_mac = EVP_MAC_fetch(NULL, "HMAC", NULL);
 
-    chimera_smb_abort_if(!ctx->hmac_mac, "Failed to fetch HMAC MAC");
+    chimera_smb2_abort_if(!ctx->hmac_mac, "Failed to fetch HMAC MAC");
 
     ctx->cmac_mac = EVP_MAC_fetch(NULL, "CMAC", NULL);
 
-    chimera_smb_abort_if(!ctx->cmac_mac, "Failed to fetch CMAC MAC");
+    chimera_smb2_abort_if(!ctx->cmac_mac, "Failed to fetch CMAC MAC");
 
     ctx->gcm = EVP_CIPHER_fetch(NULL, "AES-128-GCM", NULL);
 
-    chimera_smb_abort_if(!ctx->gcm, "Failed to fetch AES-128-GCM cipher");
+    chimera_smb2_abort_if(!ctx->gcm, "Failed to fetch AES-128-GCM cipher");
 
     return ctx;
 } /* chimera_smb_signing_ctx_new */
 
-void
+SYMBOL_EXPORT void
 chimera_smb_signing_ctx_destroy(struct chimera_smb_signing_ctx *ctx)
 {
     EVP_MAC_free(ctx->hmac_mac);
@@ -64,8 +64,8 @@ chimera_smb_signing_ctx_destroy(struct chimera_smb_signing_ctx *ctx)
  *  - For SMB 3.1.1 signing key: label = "SMBSigningKey" (no NUL), context = PreauthHash (SHA-512), 64 bytes.
  *  - For SMB 3.0/3.0.2 signing key: label = "SMB2AESCMAC\0", context = "SmbSign\0".
  */
-int
-kdf_counter_hmac_sha256_ossl3(
+SYMBOL_EXPORT int
+chimera_smb_kbkdf(
     const uint8_t *key,
     size_t         key_len,
     const void    *label,
@@ -124,7 +124,7 @@ kdf_counter_hmac_sha256_ossl3(
 
 /* Extend an SMB 3.1.1 preauth-integrity hash: hash = SHA512(hash || msg).
  * `hash` is the running 64-byte value, updated in place. */
-void
+SYMBOL_EXPORT void
 chimera_smb_preauth_extend(
     uint8_t    *hash,
     const void *msg,
@@ -133,18 +133,18 @@ chimera_smb_preauth_extend(
     EVP_MD_CTX  *md      = EVP_MD_CTX_new();
     unsigned int out_len = 0;
 
-    chimera_smb_abort_if(!md, "EVP_MD_CTX_new failed");
+    chimera_smb2_abort_if(!md, "EVP_MD_CTX_new failed");
 
     if (EVP_DigestInit_ex(md, EVP_sha512(), NULL) != 1 ||
         EVP_DigestUpdate(md, hash, SMB2_PREAUTH_HASH_SIZE) != 1 ||
         EVP_DigestUpdate(md, msg, msg_len) != 1 ||
         EVP_DigestFinal_ex(md, hash, &out_len) != 1) {
-        chimera_smb_abort("SHA-512 preauth hash update failed");
+        chimera_smb2_abort("SHA-512 preauth hash update failed");
     }
     EVP_MD_CTX_free(md);
 } /* chimera_smb_preauth_extend */
 
-int
+SYMBOL_EXPORT int
 chimera_smb_derive_signing_key(
     int            dialect,
     void          *output,
@@ -163,32 +163,33 @@ chimera_smb_derive_signing_key(
                 memcpy(output, session_key, 16);
                 return 0;
             } else {
-                chimera_smb_error("SMB2 session key length %d is not 16 bytes", session_key_len);
+                chimera_smb2_error("SMB2 session key length %d is not 16 bytes", session_key_len);
                 return -1;
             }
         case SMB2_DIALECT_3_0:
         case SMB2_DIALECT_3_0_2:
             /* 3.0 and 3.0.2 share the SP800-108 signing-key derivation
              * (label "SMB2AESCMAC", context "SmbSign"). */
-            return kdf_counter_hmac_sha256_ossl3(session_key, session_key_len,
-                                                 label30, sizeof(label30), /* includes '\0' */
-                                                 (const uint8_t *) ctx30, sizeof(ctx30), /* includes '\0' */
-                                                 output, 16);
+            return chimera_smb_kbkdf(session_key, session_key_len,
+                                     /* both lengths include the '\0' */
+                                     label30, sizeof(label30),
+                                     (const uint8_t *) ctx30, sizeof(ctx30),
+                                     output, 16);
         case SMB2_DIALECT_3_1_1:
             /* 3.1.1 binds the signing key to the preauth-integrity hash:
              * label "SMBSigningKey", context = PreauthIntegrityHashValue. */
             if (!preauth_hash) {
-                chimera_smb_error("SMB 3.1.1 signing key derivation without preauth hash");
+                chimera_smb2_error("SMB 3.1.1 signing key derivation without preauth hash");
                 return -1;
             }
             /* Per the WPTS/Windows reference, the label includes its NUL
              * terminator ("SMBSigningKey\0"); the KDF helper then inserts the
              * SP800-108 0x00 separator before the PreauthIntegrityHashValue
              * context. Pass the full sizeof (label + NUL). */
-            return kdf_counter_hmac_sha256_ossl3(session_key, session_key_len,
-                                                 label311, sizeof(label311),
-                                                 preauth_hash, SMB2_PREAUTH_HASH_SIZE,
-                                                 output, 16);
+            return chimera_smb_kbkdf(session_key, session_key_len,
+                                     label311, sizeof(label311),
+                                     preauth_hash, SMB2_PREAUTH_HASH_SIZE,
+                                     output, 16);
         default:
             return -1;
     } /* switch */
@@ -297,12 +298,12 @@ chimera_smb_request_cmac_aes_128_cbc(
 
     mctx = EVP_MAC_CTX_new(ctx->cmac_mac);
     if (!mctx) {
-        chimera_smb_error("Failed to allocate CMAC-AES-128-CBC context");
+        chimera_smb2_error("Failed to allocate CMAC-AES-128-CBC context");
         goto done;
     }
 
     if (EVP_MAC_init(mctx, key, keylen, params) != 1) {
-        chimera_smb_error("Failed to initialize CMAC-AES-128-CBC context");
+        chimera_smb2_error("Failed to initialize CMAC-AES-128-CBC context");
         goto done;
     }
 
@@ -317,7 +318,7 @@ chimera_smb_request_cmac_aes_128_cbc(
         }
 
         if (EVP_MAC_update(mctx, cursor->iov->data + cursor->offset, chunk) != 1) {
-            chimera_smb_error("Failed to update CMAC-AES-128-CBC context");
+            chimera_smb2_error("Failed to update CMAC-AES-128-CBC context");
             goto done;
         }
 
@@ -333,12 +334,12 @@ chimera_smb_request_cmac_aes_128_cbc(
     }
 
     if (left) {
-        chimera_smb_error("Left is not 0 after updating CMAC-AES-128-CBC context");
+        chimera_smb2_error("Left is not 0 after updating CMAC-AES-128-CBC context");
         goto done;
     }
 
     if (EVP_MAC_final(mctx, out_sig16, &maclen, 16) != 1) {
-        chimera_smb_error("Failed to finalize CMAC-AES-128-CBC context");
+        chimera_smb2_error("Failed to finalize CMAC-AES-128-CBC context");
         goto done;
     }
 
@@ -454,7 +455,7 @@ chimera_smb_request_gmac_aes_128(
  * the connection: HMAC-SHA256 for 2.x, AES-128-CMAC for 3.0/3.0.2, and the
  * SMB 3.1.1 negotiated signing algorithm (GMAC / CMAC / HMAC-SHA256) for 3.1.1.
  */
-static int
+SYMBOL_EXPORT int
 chimera_smb_compute_signature_alg(
     struct chimera_smb_signing_ctx *ctx,
     uint16_t                        dialect,
@@ -486,185 +487,6 @@ chimera_smb_compute_signature_alg(
     } /* switch */
 } /* chimera_smb_compute_signature_alg */
 
-static int
-chimera_smb_compute_signature(
-    struct chimera_smb_signing_ctx *ctx,
-    struct chimera_smb_conn        *conn,
-    struct smb2_header             *hdr,
-    struct evpl_iovec_cursor       *cursor,
-    int                             length,
-    const uint8_t                  *key,
-    uint8_t                        *out_sig16)
-{
-    return chimera_smb_compute_signature_alg(ctx, conn->dialect,
-                                             conn->negotiated.signing_alg,
-                                             hdr, cursor, length, key, out_sig16);
-} /* chimera_smb_compute_signature */
-
-int
-chimera_smb_verify_signature(
-    struct chimera_smb_signing_ctx *ctx,
-    struct chimera_smb_request     *request,
-    const uint8_t                  *signing_key,
-    struct evpl_iovec_cursor       *cursor,
-    int                             length)
-{
-    struct chimera_smb_conn *conn = request->compound->conn;
-    uint8_t                  signature[16];
-    uint8_t                  calculated[16];
-    char                     recv_sig[80];
-    char                     calc_sig[80];
-    int                      rc;
-
-    memcpy(&signature, &request->smb2_hdr.signature, sizeof(signature));
-    memset(request->smb2_hdr.signature, 0, sizeof(request->smb2_hdr.signature));
-
-    rc = chimera_smb_compute_signature(ctx, conn, &request->smb2_hdr, cursor,
-                                       length, signing_key, calculated);
-
-    if (unlikely(rc != 0)) {
-        chimera_smb_error("Failed to calculate signature for dialect %x", conn->dialect);
-        return rc;
-    }
-
-    /* Constant-time comparison to avoid a timing oracle on the MAC. */
-    uint8_t sig_diff = 0;
-
-    for (size_t i = 0; i < sizeof(signature); i++) {
-        sig_diff |= signature[i] ^ calculated[i];
-    }
-
-    if (unlikely(sig_diff != 0)) {
-        format_hex(recv_sig, sizeof(recv_sig), signature, sizeof(signature));
-        format_hex(calc_sig, sizeof(calc_sig), calculated, sizeof(calculated));
-        chimera_smb_error("Received signature: %s does not match calculated signature: %s", recv_sig, calc_sig);
-        return -1;
-    }
-
-    return 0;
-} /* chimera_smb_verify_signature */
-
-int
-chimera_smb_sign_compound(
-    struct chimera_smb_signing_ctx *ctx,
-    struct chimera_smb_compound    *compound,
-    struct evpl_iovec              *iov,
-    int                             niov,
-    int                             length)
-{
-    struct chimera_smb_conn           *conn = compound->conn;
-    struct chimera_smb_session_handle *session_handle;
-    struct chimera_smb_request        *request;
-    struct evpl_iovec_cursor           cursor;
-    struct smb2_header                *hdr;
-    int                                i, rc;
-    int                                left = length, payload_length;
-    uint8_t                            signature[16];
-
-    evpl_iovec_cursor_init(&cursor, iov, niov);
-
-    if (conn->protocol == EVPL_DATAGRAM_RDMACM_RC) {
-        evpl_iovec_cursor_skip(&cursor, sizeof(struct smb_direct_hdr) + 4);
-        left -= sizeof(struct smb_direct_hdr) + 4;
-    } else {
-        evpl_iovec_cursor_skip(&cursor, sizeof(struct netbios_header));
-        left -= sizeof(struct netbios_header);
-    }
-
-    for (i = 0; i < compound->num_requests && left; i++) {
-
-        request        = compound->requests[i];
-        session_handle = request->session_handle;
-
-        /* Skip requests handled asynchronously — no reply header was written
-         * for them so there is nothing to sign. */
-        if (request->status == SMB2_STATUS_PENDING) {
-            continue;
-        }
-
-        /* We know hdr is contig since we allocated it that way */
-        hdr = evpl_iovec_cursor_data(&cursor);
-
-        evpl_iovec_cursor_skip(&cursor, sizeof(struct smb2_header));
-        left -= sizeof(struct smb2_header);
-
-        if (hdr->next_command) {
-            payload_length = hdr->next_command - sizeof(struct smb2_header);
-        } else {
-            payload_length = left;
-        }
-
-        if (request->flags & CHIMERA_SMB_REQUEST_FLAG_SIGN) {
-
-            if (unlikely(!session_handle)) {
-                chimera_smb_error(
-                    "SIGN flag set but session_handle is NULL: "
-                    "cmd=0x%x msg_flags=0x%x related=%d status=0x%x req_idx=%d/%d",
-                    request->smb2_hdr.command,
-                    request->smb2_hdr.flags,
-                    !!(request->smb2_hdr.flags & SMB2_FLAGS_RELATED_OPERATIONS),
-                    request->status, i, compound->num_requests);
-                return -1;
-            }
-
-            /* A signed response MUST advertise SMB2_FLAGS_SIGNED (MS-SMB2
-             * 3.3.4.1.1), and the signature is computed over the header *with*
-             * that flag set.  The reply header inherits the request's flags, so
-             * a response to a request the client itself signed already carries
-             * the bit — but the final SESSION_SETUP response is signed by the
-             * server even though the establishing request was unsigned.  Set
-             * the flag here, before computing the MAC, so (a) the value we sign
-             * matches the bytes the client verifies and (b) a client with
-             * signing required does not treat the response as unsigned (which
-             * mishandles channel/session setup and can crash it). */
-            hdr->flags |= SMB2_FLAGS_SIGNED;
-
-            /* A SESSION_SETUP response other than a final SUCCESS (interim
-             * MORE_PROCESSING legs and errors such as a rejected channel bind)
-             * is verified by the client against its Session.SigningKey object,
-             * which carries the dialect/algorithm of the connection the
-             * session was ESTABLISHED on — not this connection's.  Sign those
-             * with the session's algorithm (Samba bug 14512; smbtorture
-             * smb2.session.bind_negative_smb3to2* receives the bind rejection
-             * on a 2.10 connection but verifies it with the 3.x session's
-             * AES-CMAC).  A final SUCCESS response is verified with the
-             * just-derived per-channel key, whose client-side object uses this
-             * connection's algorithm. */
-            if (request->smb2_hdr.command == SMB2_SESSION_SETUP &&
-                request->status != SMB2_STATUS_SUCCESS &&
-                session_handle->session &&
-                (session_handle->session->flags & CHIMERA_SMB_SESSION_AUTHORIZED)) {
-                rc = chimera_smb_compute_signature_alg(ctx,
-                                                       session_handle->session->dialect,
-                                                       session_handle->session->sign_alg,
-                                                       hdr, &cursor,
-                                                       payload_length,
-                                                       session_handle->signing_key,
-                                                       signature);
-            } else {
-                rc = chimera_smb_compute_signature(ctx, conn, hdr, &cursor,
-                                                   payload_length,
-                                                   session_handle->signing_key,
-                                                   signature);
-            }
-
-            if (unlikely(rc != 0)) {
-                chimera_smb_error("Failed to calculate signature for dialect %x", conn->dialect);
-                return rc;
-            }
-
-            memcpy(hdr->signature, signature, sizeof(signature));
-        } else {
-            evpl_iovec_cursor_skip(&cursor, payload_length);
-        }
-        left -= payload_length;
-    }
-
-    chimera_smb_abort_if(left, "Left is not 0 after signing compound");
-
-    return 0;
-} /* chimera_smb_sign_compound */
-
 /*
  * Sign a single contiguous SMB2 message in place.
  *
@@ -677,7 +499,7 @@ chimera_smb_sign_compound(
  * (STATUS_PENDING) and final responses, which are not part of the
  * normal compound reply path.
  */
-int
+SYMBOL_EXPORT int
 chimera_smb_sign_message(
     struct chimera_smb_signing_ctx *ctx,
     int                             dialect,
