+++ src/vfs/smb/CMakeLists.txt
@@ -7,14 +7,12 @@ add_library(chimera_vfs_smb SHARED
     smb_mount.c
     smb_umount.c
     smb_ops.c
-    smb_encrypt.c
-    smb_compress.c
     smb_io.c
     smb_namespace.c
     smb_ntlm.c
 )
 
-target_link_libraries(chimera_vfs_smb chimera_common evpl crypto)
+target_link_libraries(chimera_vfs_smb chimera_smb_common chimera_common evpl crypto)
 
 install(TARGETS chimera_vfs_smb DESTINATION lib)
